requirements.txt
requests==2.31.0
openai==1.12.0
scrapfly-sdk==0.8.9
beautifulsoup4==4.12.3
lxml==5.1.0
python-dotenv==1.0.1
urllib3==2.2.0

